Move login process to the listener thread
Moves the login flow to the listener thread, and also makes the listener
thread run a smol async event loop.

This feels like a better alternative than the "nanny" conditional in the
event loops of Diku/Merc, since there is no test each heartbeat for
whether or not a user has a character attached, and therefore whether to
treat them as in the login flow or not.

Also, an async function is much more easier to write a step-by-step
login flow, than trying to use an enum of possible login states. In
particular, it's always annoying that the next step's prompt is written
to the user's stream in the previous step's code branch.