Support loading players from disk
Loads a JSON file for the player name and, if found, attempts to
deserialize the character from JSON. If not found, fall back to new
character creation.

TODO:
* Consider TOML instead, for the simple reason that it contains line
breaks, which will be easier for an editor to view when it's 10,000
characters long
* Add automatic gzipping of pfiles; this may make the desire for TOML
redundant, because we'd never be viewing them in an editor anymore.