Add a primitive area/room loader
We're using TOML. Ugh **why?** Because it's more human-readable to me
than JSON. Sorry, I said it. I can visually parse a DikuMUD .are file,
and toml is pretty close to that, plus multi-line strings. Fucking deal
with it, future Bees.

Also added a separate binary just for testing the ergonomics of the
parser.