Default a player to a room print it out
Crucially, I change the Connection::write method to not actually send a
message over the connection stream, but to write to an internal byte
buffer. Then I add a second connection Connection::write_flush which
writes the buffer's contents to the stream (and then clears the buffer).

I also add a default room to new characters, to ease testing.