Interpret commands
Implements a table of commands, performs a case-insensitive lookup based
on partial matching, and executes the command by invoking the function
in the table.

The signature of each function is super gnarly, and it's only going to
get worse when objects and other things get added to the game.

It'd be lovely to pass in a single World object, but that involves
untangling lots of lifetime knots. The only reason the game works as
currently written is because characters/connections/rooms/etc. are all
separate collections with their own borrow rules. Is RefCell my savior?