Add a function to take an argument from a string
`util::take_argument(s: &str) -> Option((&str, &str))` takes a string
said to be a stringified list of arguments and returns the first
argument from that string. It is quote-aware, so it will return
multi-word args if quoted properly. It will return None when the input
string is empty. It requires all input strings to end in `'\r'` or
`'\n'`, or it will panic.