Add a take_command function like take_argument
Differs slightly from take_argument as it is not quote-aware, it only
takes the first blob of non-ascii-whitespace it sees. This is because
commands are single-word only.

Also fixes an interesting error: when a user goes linkdead (at least on
the terminal I'm using to test) the server reads an empty string as
input, and continues to read an empty string as input indefinitely.

In the previous version, the empty input string would match with the
first command ("north") because of course every string begins with the
empty string, and when it executed that command and sent output back to
the client, that would cause the connection to stop reading empty
strings as input and finally trigger an IO error, which triggered the
MUD to mark the connection as dead and remove it.

In the new version, an empty string panics when interpreted, so there is
a filter that skips command interpretation on an empty input. This
removed the erroneous command execution, which prevented the output
getting sent to the disconnected stream, which kept the stream reading
empty strings indefinitely, which prevented the connection from being
marked as linkdead. (!!!!!!)

TODO: Okay there is obviously more edge cases that I need to handle to
make my TCP streams reliable. I suspect that my solution (on empty
input, mark as linkdead just like it were an IO error) is also a bug
that just happens to "work".