Separate "player" from "character" info
The `Character` struct is going to be shared with NPCs, like Diku does
it, so I create a new struct `Player` that contains name and password
(and eventually other options), leaving `Character` with only the fields
that both player chars and NPCs will share.

`PlayerRecord` is what gets serialized/deserialized, and contains both
player and character data; it gets destructured as soon as it gets
passed from the listener thread to the game loop.

`ConnectionBuilder` is a helper struct that has fewer fields than
`Connection`. The only problem it solves is that it allows `Connection`
to change its `character` field from an optional to a concrete value
(because connections don't get assigned a character index initially).