Load NPCs and spawn them in rooms
Using the shared `Character` struct, we can deserialize NPCs from area
files, and load them into an NPCs hash map. Then, load instances of them
in the rooms they start in.

TODO: the `in-room` field is only useful for players. NPCs will want a
separate reset mechanism, maybe like Diku does it, so that multiple
copies can be loaded at once (if we want to do that)

TODO: putting NPCs in the Characters arena is not gonna work ultimately.
When an NPC dies or whatever, we won't have an Index for it saved
anywhere (like we do for player characters) in order to remove it from
the arena. *grump*

TODO: just to get it up and running, we're iterating through every
character and filtering by "in-room" field to decide who the player
can interact with. We'll *definitely* need a Vec of references or an
intrusive list per room going forward.