Impl `Write` for `Connection`
This lets me use the `write!` macro with it, instead of frequent method
calls to push output text line by line.

Not really sure what I'm supposed to do with the io errors from `write!`
though, because it's really writing to a String, so the only time it
would fail is if it failed to allocate; I guess I should crash then huh.