Add multiple keywords for NPCs
This changes the Character's name field (String) for a keywords field
(Vec<String>). For players, this will remain a single element unless
changed manually in pfiles (which you shouldn't do).

In places where we wanted to get the Character's (Player's) name, we
defer to the Player name field, which previously was just a duplicate of
the Character name field. Now, the Character keyword field is explicitly
for the purpose of interacting with a character.