Implement objects
As an experiment, trying `intrusive_collections` crate. The only linked
lists that an object gets added to is the "master" list of all objects
and the room-specific lists that describe what objects are "lying
around". Each object is inside a `Rc` so it can be in 2 lists at once.

I'm comparing to how Characters are stored: in an arena, and referenced
by index. I'm not sure which data structure I prefer more.